refactor: entity package 통일을 위한 폴더이동
