+++ src/main/java/com/letsTravel/LetsTravel/LetsTravelApplication.java
@@ -3,8 +3,10 @@ package com.letsTravel.LetsTravel;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 @EnableAspectJAutoProxy
 public class LetsTravelApplication {
