+++ src/main/java/com/letsTravel/LetsTravel/domain/profile/UserRole.java
@@ -1,4 +1,4 @@
-package com.letsTravel.LetsTravel.security.model;
+package com.letsTravel.LetsTravel.domain.profile;
 
 public enum UserRole {
     ROLE_USER, ROLE_ADMIN;
