+++ src/main/java/com/letsTravel/LetsTravel/security/config/SecurityConfig.java
@@ -1,20 +1,17 @@
 package com.letsTravel.LetsTravel.security.config;
 
 import com.letsTravel.LetsTravel.security.jwt.JwtTokenFilter;
-import com.letsTravel.LetsTravel.security.model.UserRole;
+import com.letsTravel.LetsTravel.domain.profile.UserRole;
 import com.letsTravel.LetsTravel.security.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.security.web.authentication.logout.LogoutFilter;
