+++ src/main/java/com/letsTravel/LetsTravel/security/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.letsTravel.LetsTravel.security.repository;
 
-import com.letsTravel.LetsTravel.security.model.User;
+import com.letsTravel.LetsTravel.domain.profile.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
