+++ src/main/java/com/letsTravel/LetsTravel/security/service/UserService.java
@@ -1,5 +1,7 @@
 package com.letsTravel.LetsTravel.security.service;
 
+import com.letsTravel.LetsTravel.domain.profile.User;
+import com.letsTravel.LetsTravel.domain.profile.UserRole;
 import com.letsTravel.LetsTravel.security.jwt.JwtTokenUtil;
 import com.letsTravel.LetsTravel.security.model.*;
 import com.letsTravel.LetsTravel.security.repository.UserRepository;
